Write spectrogram generator
The spectrogram is really only a visual analysis tool, and it will
mostly allow a way for the operator to get an idea of what's going on.
In this case I am the operator, and I need to know what's going on if I
am going to be writing good code.