Create window function calibration program
Evidently, neither the mean absolute value nor the root mean square of
the window function were the right values to normalize the amplitude of
a signal multiplied by the window function. Instead, we have turned to
experimental data collection.

The program added runs standalone with a single configuration file, and
is meant to give feedback on how to adjust the correction factor of the
window function.