Avoid declarations at beginning of switch case
Apparently some versions of gcc don't like when I declare a variable at
the beginning of a switch case. I understand why, and I can attest to
the fact that I didn't intend to use any of those variables anywhere
else, but I suppose it's good practice to limit scope as much as
possible.